add integration tests for dev/prod (#4446)
currently, the only difference between development and production modes is the behavior of the instanceOf() check

"production" mode is currently enabled by setting `process.env.NODE_ENV` to `production` with the default mode being "development" mode.

this PR demonstrates how to enable production mode with node, bun, deno, esbuild, rollup, swc, webpack, and rspack and demonstrates the default of development mode in those environments as well as with jest and vitest.

extracted from #4437