refactor(execution): extract withCleanup from mapAsyncIterable (#4496)
motivations:

1. separation of concerns (`mapAsyncIterable` should have single
function)
2. `withCleanUp` is useful independently, i.e. it removes the need to
"hand-roll" AsyncGenerators across the codebase

Some background for Point 2:

Built in async generators have a feature/quirk/annoyance that all of
their calls are sequential. If a call to `gen.next()` is pending, i.e.
waiting on the return of a promise (`... const result = await
fetchExternalData; yield result ... `), then the consumer cannot call
`gen.return()` manually to abort the generator, the call to
`gen.return()` will not initiate until the external fetch is awaited.
This is because Async Generators (and promises more broadly) do not have
cancellation as a built-in-feature.

Hand-rolled Async Generators can allow a call to `gen.return()` to be
used to abort the generator; they can start immediately. Technically,
they can even return out of order with that pending `gen.next()`,
although it's helpful if they always return in order, like with
Repeaters (https://repeater.js.org/).

We were previously using "hand-rolled" Async Generators for exactly this
purpose. With this change, only `withCleanUp` and `mapAsyncIterable`
stay as "hand-rolled" Async Generators, but we can use actual Async
Generators in our actual incremental delivery code (within `Incremental
Publisher` ) and subscription test code (`simplePubSub`).

This is good for two reasons:

A. Using JavaScript primitives is more modern/ergonomic
B. Hopefully will reduce churn and bug surface.

In terms of `B`, this PR also fixes a potential bug introduced when we
upgraded TS to 5.2. That version of TypeScript requires "hand-rolled"
Async Generators to include an `[Symbol.asyncDispose]` property method,
but the runtime is responsible for supplying the functionality polyfill.
When we upgraded, we by accident required our users to have that
functionality (which is present in all of our tested versions) but not
guaranteed. This change updates `mapAsyncIterable` and `withCleanUp` to
fall back to `Symbol.for("asyncDispose")` which should avoid a crash.
The broader codebase should not have to be aware of this level of detail
when using these helpers!