polish(Queue): further simplify algorithm (#4501)
=> further simplifies Queue algorithm at the expense of (potentially) a few extra micro-tasks.

The extra micro-tasks actually may lead to a few extra items being batched depending on the exact nature of the executor, potentially an added side-benefit.

Within our GraphQL code, there are no test changes, as the Queue executor just saves the push/stop methods and passes them to the IncrementalGraph.